Show and focus URL bar when creating a new tab.
This also shows and focuses the URL bar when creating a new window,
since a new window immediately creates a new tab.

I elected to allow dispatching lists (Arrays) of actions to save renders
and set a pattern for avoiding races in the future.  In spirit, this is
similar to https://twitter.com/dan_abramov/status/656074974533459968 and
https://github.com/tshelburne/redux-batched-actions.  However, this
approach works with our thunk middleware (and those don't).