Review comment 2: Store more UI state per-Page.
This doesn't separate the storage.  Lin Clark has work to do that; I'm
not going to bother to do it here, since it'll be the same for the old
userTypedLocation and the new fields that are now per-Page.

I've elected not to hide the completion results when tabs are switched;
I feel like that's a separate choice, which we may or may not
want.  (Firefox hides them when you tab away, and doesn't show them when
you tab back.)